Some high level changes
